fedors solution; very simple, but does not work for flat lines for example. need to be cleverer in removing points.

+++ src/line.jl
@@ -125,7 +125,7 @@ function prepend!(m::Line,x,y)
 end
 
 "delete an index"
-function delete!(m::Line,idx::Int)
+function delete!(m::Line,idx)
     deleteat!(m.x,idx)
     deleteat!(m.y,idx)
     reconfigure!(m)
